# docs/openapi.yaml
openapi: 3.2.0
info:
  title: Bill Payment API
  description: |
    A comprehensive RESTful API for bill payment operations, designed for API automation practice.

    ## Overview
    This API provides endpoints to manage billers, bills, payments, payment methods, users, and file uploads.
    It supports various authentication methods, content types, and includes realistic business logic.

    ## Authentication
    The API supports multiple authentication methods. All endpoints except `/health` require authentication.

    ### API Key (Recommended for Testing)
    - **Header**: `X-API-Key: demo-api-key-123`
    - **Query Parameter**: `?api_key=demo-api-key-123`

    ### Bearer Token (JWT)
    ```
    Authorization: Bearer demo-jwt-token-456
    ```

    ### Basic Authentication
    ```
    Authorization: Basic ZGVtbzpwYXNzd29yZDEyMw==
    ```
    (Credentials: `demo:password123`)

    ### Cookie-Based Session (NEW)
    ```
    Cookie: session_id=demo-session-abc123
    ```
    Session cookies are validated for authentication. Use the demo session token above for testing.

    ### OAuth2 Client Credentials
    ```bash
    POST /oauth/token
    Content-Type: application/x-www-form-urlencoded

    grant_type=client_credentials&client_id=demo-client&client_secret=demo-secret-789
    ```

    ## Content Types
    The API supports multiple content types:

    ### JSON (Default)
    - Request: `Content-Type: application/json`
    - Response: `Accept: application/json`

...

...

...

...

...

    ```
...
servers:
...
paths:
...
  /v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Register a webhook subscription
      description: |
        The `secret` is returned **only** on creation. Deliveries are signed with it
        using `X-Webhook-Signature: t=<unix>,v1=<hex HMAC-SHA256 of "<t>.<body>">`,
        which is the shape most payment providers use - so verifying it is realistic
        practice.
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items: { type: string }
                  default: ["*"]
                description: { type: string }
                isActive: { type: boolean, default: true }
            example:
              url: https://webhook.site/your-unique-id
              events: [payment.completed, payment.failed]
              description: Payment notifications
      responses:
        '201':
          description: Subscription created (includes the signing secret)
          headers:
            Location:
              schema: { type: string }
        '400':
          description: Invalid URL or unknown event name
    get:
      tags: [Webhooks]
      summary: List webhook subscriptions
      operationId: listWebhooks
      responses:
        '200':
          description: Subscription list
...
  /v1/webhooks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Webhooks]
      summary: Get a webhook subscription
      operationId: getWebhook
      responses:
        '200': { description: Subscription }
        '404': { description: Not found }
    patch:
      tags: [Webhooks]
      summary: Update a webhook subscription
      operationId: updateWebhook
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string }
                events:
                  type: array
                  items: { type: string }
                description: { type: string }
                isActive: { type: boolean }
      responses:
        '200': { description: Updated }
        '404': { description: Not found }
    delete:
      tags: [Webhooks]
      summary: Delete a subscription and its delivery log
      operationId: deleteWebhook
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }
  /v1/webhooks/{id}/test:
    post:
      tags: [Webhooks]
      summary: Fire a test delivery immediately
      description: |
        Sends one event to the subscription's URL and returns the delivery outcome,
        including the status code the endpoint returned and the signature that was
        sent. Responds `502` when the endpoint did not accept the delivery.
      operationId: testWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                event: { type: string, default: test.ping }
                data: { type: object }
      responses:
        '200':
          description: Delivered
          headers:
            X-Delivery-Id:
              schema: { type: string }
        '404': { description: Subscription not found }
        '502': { description: The endpoint rejected or did not answer the delivery }
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed in header
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: API key passed as query parameter
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication (username:password)
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: |
        Session cookie authentication. 
        Demo value: `session_id=demo-session-abc123`
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            read:all: Read access to all resources
            write:all: Write access to all resources
        password:
          tokenUrl: /oauth/token
          scopes:
            read:all: Read access
            write:all: Write access
...
security:
  - ApiKeyHeader: []
  - ApiKeyQuery: []
  - BearerAuth: []
  - BasicAuth: []
  - CookieAuth: []
  - OAuth2: []